requirements.txt
fastapi
uvicorn[standard]
scrapy
requests
python-dotenv
